+++ src/Aggregation/Traits/MissingTrait.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Elastica\Aggregation\Traits;
+
+trait MissingTrait
+{
+    /**
+     * Defines how documents that are missing a value should be treated.
+     *
+     * @param mixed $missing
+     *
+     * @return $this
+     */
+    public function setMissing($missing): self
+    {
+        return $this->setParam('missing', $missing);
+    }
+}