+++ tests/Aggregation/AvgTest.php
@@ -30,6 +30,25 @@ class AvgTest extends BaseAggregationTest
         $this->assertEquals((5 + 8 + 1 + 3) / 4.0, $results['avg']['value']);
     }
 
+    /**
+     * @group functional
+     */
+    public function testAvgAggregationWithMissingValue(): void
+    {
+        $agg = new Avg('avg');
+        $agg->setField('price');
+        $agg->setMissing(72);
+
+        $query = new Query();
+        $query->addAggregation($agg);
+
+        $resultSet = $this->_getIndexForTest()->search($query);
+        $results = $resultSet->getAggregations();
+
+        $this->assertTrue($resultSet->hasAggregations());
+        $this->assertEquals((5 + 8 + 1 + 3 + 72) / 5.0, $results['avg']['value']);
+    }
+
     protected function _getIndexForTest(): Index
     {
         $index = $this->_createIndex();
@@ -39,6 +58,7 @@ class AvgTest extends BaseAggregationTest
             new Document(2, ['price' => 8]),
             new Document(3, ['price' => 1]),
             new Document(4, ['price' => 3]),
+            new Document(5, ['anything' => 'anything']),
         ]);
 
         $index->refresh();