+++ tests/Aggregation/SumTest.php
@@ -27,6 +27,22 @@ class SumTest extends BaseAggregationTest
         $this->assertEquals(5 + 8 + 1 + 3, $results['value']);
     }
 
+    /**
+     * @group functional
+     */
+    public function testSumAggregationWithMissing(): void
+    {
+        $agg = new Sum('sum');
+        $agg->setField('price');
+        $agg->setMissing(10);
+
+        $query = new Query();
+        $query->addAggregation($agg);
+        $results = $this->_getIndexForTest()->search($query)->getAggregation('sum');
+
+        $this->assertEquals(5 + 8 + 1 + 3 + 10, $results['value']);
+    }
+
     protected function _getIndexForTest(): Index
     {
         $index = $this->_createIndex();
@@ -36,6 +52,7 @@ class SumTest extends BaseAggregationTest
             new Document(2, ['price' => 8]),
             new Document(3, ['price' => 1]),
             new Document(4, ['price' => 3]),
+            new Document(5, ['anything' => 'anything']),
         ]);
 
         $index->refresh();